requirements.txt
flask>=3.0
flask_sqlalchemy>=3.1
flask_migrate>=4.0
flask_login>=0.6
python-dotenv>=1.0
email-validator>=2.1
wtforms>=3.1
werkzeug>=3.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
